renamed directory and removed dead code
